Moved abstract cross sections out of reaction_data.hpp
