+++ src/reactions_lib/cross_sections/constant_rate_cs.hpp
@@ -1,6 +1,6 @@
 #ifndef REACTIONS_CONSTANT_RATE_CS_H
 #define REACTIONS_CONSTANT_RATE_CS_H
-#include "../reaction_data.hpp"
+#include "../cross_section_abstract.hpp"
 #include <limits>
 #include <neso_particles.hpp>
